// frontend/src/generated/api-schema.d.ts
export interface components {
    schemas: {
// ...
        MediaSummary: {
            type?: string;
            provider?: string;
            externalId?: string;
            title?: string;
            /** Format: int32 */
            year?: number;
            genre?: string;
            synopsis?: string;
            coverUrl?: string;
            externalRating?: number;
            inLibrary?: boolean;
            /** Format: int64 */
            entryId?: number;
        };
        SearchResult: {
            content?: components["schemas"]["MediaSummary"][];
            /** Format: int32 */
            page?: number;
            /** Format: int32 */
            size?: number;
            /** Format: int64 */
            totalElements?: number;
            /** Format: int32 */
            totalPages?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}